Removed settings and debugged default logging level info
The settings.py are not required anymore with the introduction of vault.
It remained in the branch due to a unsuccessful merge.

Solved bug environmental variable LOG_LEVEL will always exist but it's
value is a empty string not None as presumed in if statement of setting
FireML log level.